service/requirements.txt
amqp==2.6.1
anyjson==0.3.3
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==4.4.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==39.0.0
Django==2.2
django-celery==3.3.1
django-celery-beat==2.0.0
django-cors-headers==3.13.0
django-filter==22.1
django-ranged-response==0.2.0
django-redis==5.2.0
django-rest-knox==4.2.0
django-simple-captcha==0.5.17
django-timezone-field==4.2.3
djangorestframework==3.14.0
importlib-metadata==5.0.0
kombu==4.6.11
Markdown==3.4.1
Pillow==9.3.0
prompt-toolkit==3.0.36
psycopg2==2.8.6
pycparser==2.21
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.6
redis==4.4.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
vine==1.3.0
wcwidth==0.2.6
zipp==3.10.0
